+++ CHANGELOG.md
@@ -3,10 +3,23 @@
 `relude` uses [PVP Versioning][1].
 The changelog is available [on GitHub][2].
 
-## Unreleased: 0.6.0
+## Unreleased: 0.6.0.0
+
+* [#167](https://github.com/kowainik/relude/issues/167):
+  Rename functions `prec`/`prev`, `dupe`/`dup`.
 * [#155](https://github.com/kowainik/relude/issues/155):
   Implement `Relude.Extra.Foldable` module.
-* Use `foldr` instead of explicit recursion and `toList`.
+* Re-export `GHC.Float.atan2`.
+* [#172](https://github.com/kowainik/relude/issues/172):
+  Add `Monoid` and `Semigroup` instances for `Validation` type
+* [#177](https://github.com/kowainik/relude/issues/177):
+  Improve usage of performance pragmas.
+* [#178](https://github.com/kowainik/relude/issues/178):
+  Made `die` be polymorphic in its return type.
+* [#164](https://github.com/kowainik/relude/issues/164):
+  Reexport `ShortByteString`, `toShort`/`fromShort` functions.
+* [#162](https://github.com/kowainik/relude/pull/162):
+  Use `foldr` instead of explicit recursion and `toList`.
 
 ## 0.5.0 — Mar 18, 2019